Rename task serializer and use base serializer mixins
